+++ examples/marine/flowline/createSetup_flowline.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# Copyright (C) 2011, 2013, 2014, 2016, 2018 Torsten Albrecht and Moritz Huetten
+# Copyright (C) 2011, 2013, 2014, 2016, 2018, 2021 Torsten Albrecht and Moritz Huetten
 
 # ./createSetup_flowline.py -a 0.0 -r 10.0
 
@@ -56,8 +56,7 @@ command-line options. (We try to use command-line options whenever we can.)
 filename = "flowline_config.nc"
 nc = NC(filename, 'w', format="NETCDF3_CLASSIC")
 var = nc.createVariable("pism_overrides", 'i')
-attrs = {"ocean.always_grounded": "no",
-         "geometry.update.use_basal_melt_rate": "no",
+attrs = {"geometry.update.use_basal_melt_rate": "no",
          "stress_balance.ssa.compute_surface_gradient_inward": "no",
          "flow_law.isothermal_Glen.ice_softness": (B0) ** -3,
          "constants.ice.density": rho_ice,