+++ examples/mismip/mismip2d/run.py
@@ -146,8 +146,7 @@ class Experiment:
 
         var = nc.createVariable("pism_overrides", 'i')
 
-        attrs = {"ocean.always_grounded": "no",
-                 "geometry.update.use_basal_melt_rate": "no",
+        attrs = {"geometry.update.use_basal_melt_rate": "no",
                  "stress_balance.ssa.compute_surface_gradient_inward": "no",
                  "flow_law.isothermal_Glen.ice_softness": MISMIP.A(self.experiment, step),
                  "constants.ice.density": MISMIP.rho_i(),