+++ src/stressbalance/ShallowStressBalance.hh
@@ -38,7 +38,7 @@ namespace stressbalance {
  * Vertically-averaged ocean pressure difference at the calving front, used in the implementation of
  * the stress boundary condition at the calving front in SSA stress balance solvers.
  */
-double margin_pressure_difference(bool shelf, bool dry_mode, double H, double bed, double sea_level,
+double margin_pressure_difference(bool shelf, double H, double bed, double sea_level,
                                   double rho_ice, double rho_ocean, double g);
 
 class Inputs;